scoped enum's underlying types that are small or are bitfields changed to uint*_t; Added/Implemented Bitwise op= and op in terms of op=;

+++ Engine/Code/Engine/RHI/RHIDeviceContext.hpp
@@ -21,8 +21,8 @@ class RasterState;
 class BlendState;
 class Material;
 
-enum class BufferType : unsigned int;
-enum class PrimitiveType : unsigned int;
+enum class BufferType : uint8_t;
+enum class PrimitiveType : uint32_t;
 
 class RHIDeviceContext {
 public:
